fix: satisfy backend-size while keeping pkg-file-count green
Move operator-default runtime application into pkg/config/operatordefaultsruntime,
add LoadedFactoryConfig.MutateWorkers, and use backendsizecheck ignore directives
for consolidated CLI tests instead of inflating runtime_config.go.
