+++ pkg/cli/root_run_test.go
@@ -1,3 +1,4 @@
+// backendsizecheck:ignore-file consolidated root run and operator-default CLI tests remain together until dedicated CLI test seams split.
 package cli
 
 import (
